# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"OracleInfo": {
				"properties": {
					"publicKey": {
						"type": "string"
					},
					"availableTickers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"publicKey",
					"availableTickers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OracleAttestation": {
				"properties": {
					"timestamp": {
						"type": "string"
					},
					"lastPrice": {
						"type": "string"
					},
					"attestation": {
						"properties": {
							"messageHash": {
								"type": "string"
							},
							"message": {
								"type": "string"
							},
							"signature": {
								"type": "string"
							}
						},
						"required": [
							"messageHash",
							"message",
							"signature"
						],
						"type": "object"
					}
				},
				"required": [
					"timestamp",
					"lastPrice",
					"attestation"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "oracle",
		"version": "0.0.0",
		"description": "Reference implementation for an oracle producing price feed cryptographic attestation to be consumed by Elements TapScript opcode CHECKSIGFROMSTACK",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/oracle": {
			"get": {
				"operationId": "GetInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OracleInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Oracle"
				],
				"security": [],
				"parameters": []
			}
		},
		"/oracle/{ticker}": {
			"get": {
				"operationId": "GetAttestationForTicker",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/OracleAttestation"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Oracle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "ticker",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}